+++ app/src/main/java/com/example/findu/presentation/ui/my/adapter/MyInterestRvAdapter.kt
@@ -13,8 +13,8 @@ import com.example.findu.presentation.model.MyRecentHistoryRv
 import com.example.findu.presentation.type.AnimalStateType
 
 class MyInterestRvAdapter(
-    private val onKeepClick: (Int, Boolean, String) -> Unit,
-    private val onItemClick: (Int) -> Unit = {}
+    private val onKeepClick: (Long, Boolean, String) -> Unit,
+    private val onItemClick: (Long) -> Unit = {}
 ) : ListAdapter<MyInterestRv, MyInterestRvAdapter.MyListRvViewHolder>(diffUtil) {
 
     inner class MyListRvViewHolder(private val binding: ItemMyHistoryBinding) :
